Remove test page
