+++ samples/CommunityToolkit.Maui.Sample/Pages/Layouts/UniformItemsLayoutPage.xaml
@@ -4,38 +4,49 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:pages="clr-namespace:CommunityToolkit.Maui.Sample.Pages"
     xmlns:mct="clr-namespace:CommunityToolkit.Maui.Layouts;assembly=CommunityToolkit.Maui"
-    x:Class="CommunityToolkit.Maui.Sample.Pages.Layouts.UniformItemsLayoutPage">
-    <VerticalStackLayout Padding="{StaticResource ContentPadding}">
+    xmlns:layouts="clr-namespace:CommunityToolkit.Maui.Sample.ViewModels.Layouts"
+    x:Class="CommunityToolkit.Maui.Sample.Pages.Layouts.UniformItemsLayoutPage"
+    x:TypeArguments="layouts:UniformItemsLayoutViewModel"
+    x:DataType="layouts:UniformItemsLayoutViewModel">
 
-        <Label Text="The UniformItemsLayout creates a grid where each Row and Column will have the same size" 
-                Margin="0,0,0,12"/>
+    <ScrollView Padding="0">
+        <VerticalStackLayout Spacing="12" >
+            <Label Text="The UniformItemsLayout creates a grid where each Row and Column will have the same size"
+                       Margin="0,0,0,12"/>
 
-        <Label Text="UniformItemsLayout with default settings"/>
-        <mct:UniformItemsLayout>
-            <BoxView BackgroundColor="Red" WidthRequest="25" HeightRequest="25"/>
-            <BoxView BackgroundColor="Green" WidthRequest="25" HeightRequest="25"/>
-            <BoxView BackgroundColor="Yellow" WidthRequest="25" HeightRequest="25"/>
-        </mct:UniformItemsLayout>
-        
-        <Label Text="UniformItemsLayout with MaxColumns=1 and MaxRows=1"/>
-        <mct:UniformItemsLayout MaxColumns="1" MaxRows="1">
-            <BoxView BackgroundColor="Red" WidthRequest="25" HeightRequest="25"/>
-            <BoxView BackgroundColor="Green" WidthRequest="25" HeightRequest="25"/>
-            <BoxView BackgroundColor="Yellow" WidthRequest="25" HeightRequest="25"/>
-        </mct:UniformItemsLayout>
-        
-        <Label Text="UniformItemsLayout with MaxRows=1"/>
-        <mct:UniformItemsLayout MaxRows="1">
-            <BoxView BackgroundColor="Red" WidthRequest="25" HeightRequest="25"/>
-            <BoxView BackgroundColor="Green" WidthRequest="25" HeightRequest="25"/>
-            <BoxView BackgroundColor="Yellow" WidthRequest="25" HeightRequest="25"/>
-        </mct:UniformItemsLayout>
+            <Button Text="Add Color Box To Grids" Clicked="HandleAddButtonClicked" HorizontalOptions="Center" VerticalOptions="Center" Padding="8"/>
 
-        <Label Text="UniformItemsLayout with MaxColumns=1"/>
-        <mct:UniformItemsLayout MaxColumns="1">
-            <BoxView BackgroundColor="Red" WidthRequest="25" HeightRequest="25"/>
-            <BoxView BackgroundColor="Green" WidthRequest="25" HeightRequest="25"/>
-            <BoxView BackgroundColor="Yellow" WidthRequest="25" HeightRequest="25"/>
-        </mct:UniformItemsLayout>
-    </VerticalStackLayout>
+            <Label Text="UniformItemsLayout with default settings"/>
+            <mct:UniformItemsLayout x:Name="UniformItemsLayout_Default">
+                <BoxView BackgroundColor="Black" HeightRequest="25" WidthRequest="25"/>
+                <BoxView BackgroundColor="Green" HeightRequest="25" WidthRequest="25"/>
+                <BoxView BackgroundColor="Purple" HeightRequest="25" WidthRequest="25"/>
+                <BoxView BackgroundColor="Orange" HeightRequest="25" WidthRequest="25"/>
+            </mct:UniformItemsLayout>
+
+            <Label Text="UniformItemsLayout with MaxColumns=2 and MaxRows=2"/>
+            <mct:UniformItemsLayout x:Name="UniformItemsLayout_MaxRows2MaxColumns2" MaxColumns="2" MaxRows="2">
+                <BoxView BackgroundColor="Black" HeightRequest="25" WidthRequest="25"/>
+                <BoxView BackgroundColor="Green" HeightRequest="25" WidthRequest="25"/>
+                <BoxView BackgroundColor="Purple" HeightRequest="25" WidthRequest="25"/>
+                <BoxView BackgroundColor="Orange" HeightRequest="25" WidthRequest="25"/>
+            </mct:UniformItemsLayout>
+
+            <Label Text="UniformItemsLayout with MaxRows=1"/>
+            <mct:UniformItemsLayout x:Name="UniformItemsLayout_MaxRows1" MaxRows="1">
+                <BoxView BackgroundColor="Black" HeightRequest="25" WidthRequest="25"/>
+                <BoxView BackgroundColor="Green" HeightRequest="25" WidthRequest="25"/>
+                <BoxView BackgroundColor="Purple" HeightRequest="25" WidthRequest="25"/>
+                <BoxView BackgroundColor="Orange" HeightRequest="25" WidthRequest="25"/>
+            </mct:UniformItemsLayout>
+
+            <Label Text="UniformItemsLayout with MaxColumns=1"/>
+            <mct:UniformItemsLayout x:Name="UniformItemsLayout_MaxColumns1" MaxColumns="1">
+                <BoxView BackgroundColor="Black" HeightRequest="25" WidthRequest="25"/>
+                <BoxView BackgroundColor="Green" HeightRequest="25" WidthRequest="25"/>
+                <BoxView BackgroundColor="Purple" HeightRequest="25" WidthRequest="25"/>
+                <BoxView BackgroundColor="Orange" HeightRequest="25" WidthRequest="25"/>
+            </mct:UniformItemsLayout>
+        </VerticalStackLayout>
+    </ScrollView>
 </pages:BasePage>
