+++ samples/CommunityToolkit.Maui.Sample/ViewModels/Layouts/LayoutsGalleryViewModel.cs
@@ -10,8 +10,6 @@ public class LayoutsGalleryViewModel : BaseGalleryViewModel
 		{
 			SectionModel.Create<UniformItemsLayoutViewModel>(nameof(UniformItemsLayoutPage),
 				"A Grid where all rows and columns have the same size"),
-			SectionModel.Create<UniformItemsLayoutTestViewModel>(nameof(UniformItemsLayoutTestPage),
-				"A test page that adds new views to UniformItemsLayout"),
 		})
     {
     }
