--- samples/CommunityToolkit.Maui.Sample/ViewModels/Layouts/UniformItemsLayoutTestViewModel.cs
@@ -1,6 +1,0 @@
-﻿namespace CommunityToolkit.Maui.Sample.ViewModels.Layouts;
-
-public class UniformItemsLayoutTestViewModel : BaseViewModel
-{
-
-}
